+++ src/extension/app/plugins/publish/publish.js
@@ -38,6 +38,7 @@ export function createPublishPlugin(appStore) {
         if (res) {
           const actionCallback = () => {
             appStore.switchEnv('prod', newTab(evt));
+            appStore.closeToast();
           };
 
           const closeCallback = () => {