+++ src/extension/app/plugins/unpublish/unpublish.js
@@ -65,6 +65,7 @@ export function createUnpublishPlugin(appStore) {
           if (res) {
             const actionCallback = () => {
               appStore.reloadPage(newTab(evt));
+              appStore.closeToast();
             };
 
             const closeCallback = () => {